% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_hoeffding}
\alias{max_hoeffding}
\title{Maximum data size for Hoeffding's statistics}
\usage{
max_hoeffding()
}
\value{
100413509 in 128-bit mode,
14081 in 64-bit mode.
}
\description{
Data at most this size should not cause an overflow in the computation
of Hoeffding's D statistic or its refinement. This may depend on the
availability of 64-bit or 128-bit words to the compiler in use.
}
\examples{
max_hoeffding()
}
\seealso{
\code{\link{hoeffding.D.test}}
\code{\link{hoeffding.refined.test}}
}
