% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_taustar}
\alias{max_taustar}
\title{Maximum data size for the T* statistic}
\usage{
max_taustar()
}
\value{
6721987087 in 128-bit mode,
102569 in 64-bit mode.
}
\description{
Data at most this size should not cause an overflow in the computation
of Tau*. This may depend on the availability of 64-bit or 128-bit words
to the compiler in use.
}
\examples{
max_taustar()
}
\seealso{
\code{\link{tau.star.test}}
}
